feat: new root api
